Fix CPU benchmark
In PHP 7.4 passing invalid characters will now generate a deprecation notice.

This is because we don't have how to convert a binary number to a decimal by passing a decimal number to the function. I believe it was a mistake. That's why I corrected the function that converts a decimal number to a binary number.

https://www.php.net/manual/en/function.bindec.php#refsect1-function.bindec-changelog